refactor move validator to validator/mod.rs
